Add config.yml

+++ Vagrantfile
@@ -1,3 +1,9 @@
+require 'yaml'
+
+current_dir   = File.dirname(File.expand_path(__FILE__))
+configs       = YAML.load_file("#{current_dir}/config.yml")
+custom_config = configs['configs']
+
 Vagrant.configure(2) do |config|
 
 	config.vm.box = "loispuig/debian-jessie-amd64"
@@ -5,8 +11,8 @@ Vagrant.configure(2) do |config|
 	# Create a forwarded port mapping which allows access to a specific port
 	# within the machine from a port on the host machine. In the example below,
 	# accessing "localhost:8080" will access port 80 on the guest machine.
-	#config.vm.network "forwarded_port", guest: 80, host: 8080
-	#config.vm.network "forwarded_port", guest: 443, host: 8443
+	config.vm.network "forwarded_port", guest: 80, host: 8080
+	config.vm.network "forwarded_port", guest: 443, host: 8443
 
 	# Create a private network, allows host-only access to the machine using a specific IP.
 	config.vm.network "private_network", ip: "10.10.10.10"
@@ -26,29 +32,60 @@ Vagrant.configure(2) do |config|
 	config.vm.synced_folder "log/php",       "/var/log/php",      create: true, owner: 'vagrant',  group: 'vagrant', mount_options: ["dmode=777,fmode=666"]
 	config.vm.synced_folder "log/apache2",   "/var/log/apache2",  create: true, owner: 'vagrant',  group: 'vagrant', mount_options: ["dmode=777,fmode=666"]
 	config.vm.synced_folder "log/redis",     "/var/log/redis",    create: true, owner: 'vagrant',  group: 'vagrant', mount_options: ["dmode=777,fmode=666"]
-	config.vm.synced_folder "www/",          "/var/www",          create: true, owner: 'www-data', group: 'vagrant', mount_options: ["dmode=775,fmode=664"]
+	config.vm.synced_folder "letsencrypt",   "/etc/letsencrypt",  create: true
+	config.vm.synced_folder "www/",          "/var/www",          create: true, owner: 'www-data', group: 'vagrant', mount_options: ["dmode=775,fmode=774"]
 
 	config.vm.provider "virtualbox" do |vb|
-		vb.name = "vagrant-dhdp"
+		vb.name = "vagrant-"+custom_config['hostname']
 		vb.gui = false
-		vb.memory = 1024
-		vb.cpus = 1
+		vb.memory = 2048
+		vb.cpus = 2
 		# No matter how much CPU is used in the VM,
 		# no more than 50% would be used on your own host machine
-		vb.customize ["modifyvm", :id, "--cpuexecutioncap", "50"]
+		#vb.customize ["modifyvm", :id, "--cpuexecutioncap", "50"]
 	end
+
+	# Warns to use Bindler
+	unless Vagrant.has_plugin?("vagrant-hostmanager")
+		puts "--- WARNING ---"
+		puts "--- Plugin vagrant-hostmanager not installed, use command : vagrant plugin install vagrant-hostmanager"
+	end
+
+	if Vagrant.has_plugin?("vagrant-hostmanager")
+		config.hostmanager.enabled = true
+		config.hostmanager.manage_host = true
+		config.hostmanager.ignore_private_ip = false
+		config.hostmanager.include_offline = true
+		config.vm.define 'vagrant-dhdp' do |node|
+			node.vm.hostname = custom_config['hostname']+'.'+custom_config['domain']
+			#node.vm.network :private_network, ip: '10.10.10.10'
+			node.hostmanager.aliases = custom_config['aliases']
+		end
+  	end
+
+  	# hostmanager provisioner
+	config.vm.provision :hostmanager
 
 	config.vm.provision "shell", inline: <<-SHELL
 		cd /vagrant/puppet
 		rm -Rf log/mysql/* log/php/* log/apache2/* log/redis/*
-		rm -Rf .tmp .librarian Puppetfile.lock
-		librarian-puppet install --verbose --clean
+		LIBRARIANRESET=0
+		if [ ! -d .librarian -o ${LIBRARIANRESET} -eq 1 ] ; then
+			rm -Rf .tmp .librarian Puppetfile.lock
+			librarian-puppet install --clean --verbose
+		#else
+			#librarian-puppet outdated | cut -d ' ' -f 1 | xargs librarian-puppet update $1 --verbose
+			#librarian-puppet update puppet-hdp --verbose
+		fi
 	SHELL
 
 	config.vm.provision "puppet" do |puppet|
 		puppet.manifests_path = "puppet"
 		puppet.manifest_file = "init.pp"
 		puppet.module_path = "puppet/modules"
-		puppet.options = "--debug --verbose"
+		puppet.options = "--debug --verbose --no-stringify_facts --trusted_node_data"
+		puppet.facter = {
+			"custom_config" => custom_config
+        }
 	end
 end
